@@ -23,16 +23,16 @@
     }
   ],
   "require": {
-    "php": "^7.2|^8.0",
+    "php": "^8.1",
     "ext-json": "*",
     "ext-mbstring": "*",
-    "monolog/monolog": "^2.0"
+    "monolog/monolog": "^3.0"
   },
   "require-dev": {
     "ext-curl": "*",
     "squizlabs/php_codesniffer": "^3.2",
-    "phpunit/phpunit": "^8.0|^9.0",
-    "vlucas/phpdotenv": "^3.3"
+    "phpunit/phpunit": "^9.0",
+    "vlucas/phpdotenv": "^5.0"
   },
   "suggest": {
     "ext-curl": "cURL generally works better and is recommended"
