@@ -3,58 +3,76 @@
 namespace jacklul\MonologTelegramHandler\Tests;
 
 use Dotenv\Dotenv;
+use Monolog\Handler\BufferHandler;
+use Monolog\Level;
 use Monolog\Logger;
 use PHPUnit\Framework\TestCase;
 use jacklul\MonologTelegramHandler\TelegramHandler;
 
 class TelegramHandlerTest extends TestCase
 {
-    private $token;
-    private $chat_id;
+    private ?string $token;
+    private ?int $chat_id;
 
-    public function __construct($name = null, array $data = array(), $dataName = '')
+    public function __construct(?string $name = null, array $data = [], $dataName = '')
     {
         if (class_exists(Dotenv::class) && file_exists(dirname(__DIR__) . '/.env')) {
-            $env = Dotenv::create(dirname(__DIR__));
+            $env = Dotenv::createImmutable(dirname(__DIR__));
             $env->load();
         }
 
-        $this->token = getenv('TELEGRAM_TOKEN');
-        $this->chat_id = getenv('TELEGRAM_CHAT_ID');
+        $this->token = $_ENV['TELEGRAM_TOKEN'] ?? getenv('TELEGRAM_TOKEN');
+        $this->chat_id = (int) ($_ENV['TELEGRAM_CHAT_ID'] ?? getenv('TELEGRAM_CHAT_ID'));
 
         parent::__construct($name, $data, $dataName);
     }
 
-    public function testWithValidArguments()
+    public function testWithValidArguments(): void
     {
         if (empty($this->token) || empty($this->chat_id)) {
             $this->markTestSkipped('Either token or chat id was not provided');
         }
 
+        $logger = new Logger('PHPUnit');
+        $handler = new TelegramHandler($this->token, $this->chat_id, Level::Debug, true, false);
+        $logger->pushHandler($handler);
+
+        try {
+            $logger->debug('PHPUnit - test with valid arguments: ' . PHP_EOL . str_repeat('-', 4096));
+        } catch (\Exception $e) {
+            $this->fail('Exception was thrown: ' . $e->getMessage());
+        }
+
+        sleep(1);
+        
         if (extension_loaded('curl')) {
             $logger = new Logger('PHPUnit');
-            $handler = new TelegramHandler($this->token, $this->chat_id, Logger::DEBUG, true, true, 5, true);
+            $handler = new TelegramHandler($this->token, $this->chat_id, Level::Debug);
             $logger->pushHandler($handler);
 
             try {
-                $logger->debug('PHPUnit - test with valid arguments and large message: ' . str_repeat('-', 4096));
+                $logger->debug('PHPUnit - test with valid arguments using cURL: ' . PHP_EOL . str_repeat('-', 4096));
             } catch (\Exception $e) {
                 $this->fail('Exception was thrown: ' . $e->getMessage());
             }
 
             $this->assertTrue(true);
+            
+            sleep(1);
         }
 
+        $this->assertTrue(true);
+        
         $logger = new Logger('PHPUnit');
-        $handler = new TelegramHandler($this->token, $this->chat_id, Logger::DEBUG, true, false, 5, (float)PHP_VERSION >= 5.6); // PHP 5.5 has trouble accessing SSL-protected URLs so the verification must be force-disabled
+        $handler = new TelegramHandler($this->token, $this->chat_id, Level::Debug);
+        $handler = new BufferHandler($handler);
         $logger->pushHandler($handler);
 
         try {
-            if (isset($result) && $result) {
-                $logger->debug('PHPUnit - test with valid arguments');
-            } else {
-                $logger->debug('PHPUnit - test with valid arguments and large message: ' . str_repeat('-', 4096));
-            }
+            $logger->debug('PHPUnit - Batch processing test 1');
+            $logger->debug('PHPUnit - Batch processing test 2');
+            $logger->debug('PHPUnit - Batch processing test 3');
+            $handler->close();
         } catch (\Exception $e) {
             $this->fail('Exception was thrown: ' . $e->getMessage());
         }
@@ -62,33 +80,29 @@ class TelegramHandlerTest extends TestCase
         $this->assertTrue(true);
     }
 
-    public function testWithInvalidToken()
+    public function testWithInvalidToken(): void
     {
         sleep(1);
 
-        if (method_exists($this, 'expectException') && method_exists($this, 'expectExceptionMessage')) {
-            $this->expectException(\RuntimeException::class);
-            $this->expectExceptionMessage('Not Found');
-        } elseif (method_exists($this, 'setExpectedException')) {
-            $this->setExpectedException(\RuntimeException::class, 'Not Found');
-        }
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Not Found');
 
         if (extension_loaded('curl')) {
             $logger = new Logger('PHPUnit');
-            $handler = new TelegramHandler('token', $this->chat_id, Logger::DEBUG, true, true, 5, true);
+            $handler = new TelegramHandler('token', $this->chat_id, Level::Debug);
             $logger->pushHandler($handler);
 
             $logger->debug('PHPUnit');
         }
 
         $logger = new Logger('PHPUnit');
-        $handler = new TelegramHandler('token', $this->chat_id, Logger::DEBUG, true, false, 5, (float)PHP_VERSION >= 5.6);
+        $handler = new TelegramHandler('token', $this->chat_id, Level::Debug, true, false);
         $logger->pushHandler($handler);
 
         $logger->debug('PHPUnit');
     }
 
-    public function testWithInvalidChatId()
+    public function testWithInvalidChatId(): void
     {
         sleep(1);
 
@@ -96,23 +110,19 @@ class TelegramHandlerTest extends TestCase
             $this->markTestSkipped('Token was not provided');
         }
 
-        if (method_exists($this, 'expectException') && method_exists($this, 'expectExceptionMessage')) {
-            $this->expectException(\RuntimeException::class);
-            $this->expectExceptionMessage('Bad Request');
-        } elseif (method_exists($this, 'setExpectedException')) {
-            $this->setExpectedException(\RuntimeException::class, 'Bad Request');
-        }
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Bad Request');
 
         if (extension_loaded('curl')) {
             $logger = new Logger('PHPUnit');
-            $handler = new TelegramHandler($this->token, 123, Logger::DEBUG, true, true, 5, true);
+            $handler = new TelegramHandler($this->token, 123, Level::Debug);
             $logger->pushHandler($handler);
 
             $logger->debug('PHPUnit');
         }
 
         $logger = new Logger('PHPUnit');
-        $handler = new TelegramHandler($this->token, 123, Logger::DEBUG, true, false, 5, (float)PHP_VERSION >= 5.6);
+        $handler = new TelegramHandler($this->token, 123, Level::Debug, true, false);
         $logger->pushHandler($handler);
 
         $logger->debug('PHPUnit');
